requirements.txt
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
matplotlib==3.7.1
pyyaml